requirements.txt
torch>=2.0.0
transformers>=4.40.0
datasets>=2.16.0
peft>=0.7.0
accelerate>=1.1.0
pyyaml>=6.0
Pillow>=10.0.0
tqdm>=4.66.0
